+++ srcs/put_pixel.c
@@ -1,6 +1,6 @@
 #include "tft_ili9486_at2560.h"
 
-int put_pixel(uint16_t sx, uint16_t ex, uint16_t sy, uint16_t ey, uint32_t rgb) {
+int put_pixel(uint16_t sx, uint16_t ex, uint16_t sy, uint16_t ey, uint16_t rgb) {
 	if (sx > ex || ex >= 320 || sy > ey || ey >= 480) {
 		return 0;
 	}
@@ -20,6 +20,7 @@ int put_pixel(uint16_t sx, uint16_t ex, uint16_t sy, uint16_t ey, uint32_t rgb) 
 	write_bus(ey & 0x00FF, INDEX);
 
 	// Memory Write
+	// 11 1111 0000 0000 0000
 	write_bus(0x2C, COMMAND);
 	for (uint16_t y = sy; y <= ey; ++y) {
 		for (uint16_t x = sx; x <= ex; ++x) {
